Updates to initialization code
ADC Initialization was implemented and successfully tested.
A calibration of the unit is performed and the Successfull/Not Successful is determined based off of the cal results.
The ISR has not yet been integrated.
Frequency of PWM signal was changed to 75Hz.  This allows for more varience in the PW.
SPI has been initialized.  Nothing has been tested yet, this still needs to be done.  Interrupts need to be set up.