SPI Bus to IMU Working
ADC Interrupt Setup

SPI transmitting to IMU verified.

Self Test added where SPI will write to IMU "WHO_AM_I" register. If the correct response is heard back, it will mark it as a pass.

Right now requests are being made for the IMU data, but nothing is being recorded.

The IMU read function is set up to read from either of the two chips.  The chip select is an input to the function.

The IMU that we currently have was wired as an IMU0 and then as an IMU1.  It was verified that we are able to talk to it in either configuration.