Updates to SD Card Logging, RTC logging
Version 7:

Added function to get the current time from the RTC.  It populates an array as [sec, min, hr, day, month, year]

Added Data Write function to the SD Card

Added New line write function to the data card.

Right now I have it working where it prints all of the data on one line with one timestamp for all 6 IMU data pieces.

Each set of 6 reads/write takes approximately 60ms, so one time stamp is fine for this.

Add timer to trigger an interrupt every second.  This says that its time to get the pressure and temp functions.  Inside this loop the Heartbeat function is called which flashes the LED.