Integrated Motor Algorithim
Added functions to turn motor on/off
Started reaction wheel algorithim.  Only going to do LSB conversions for Z-Gyro when the reaction wheel is running.  Otherwise its using recources that we don't need to use.  Data is captured and then converted to rev/min.
Took out the two's compliment function.  Now when the data comes back from the IMU it's being placed into a 16 bit short instead of a 32 bit int.  Because it's in the 16 bit short the compiler recognizes it as a negative number, which makes no need for two's compliment.
A button will be used in order to determine if a calibration of the IMU is to be done.  This button needs to be held down on startup in order for the calibration to take place.  This is done so that a calibration does not get accidentally triggered during the flight.
Calibration is stored in flash memory.  If the button is not pressed it does not perform the calibration
The calibration is only done for the gyroscope.
The motor is working about 90% of the way.  The last issue is that the integral won't totally settle to 0 after it is spun fast in one direction and then not at all in the other direction.  This should fix itself once it is connected to the actual motor.  This is because it will spin it slightly in one direction and that will drive the integral to zero.  This will be tested once we can mount everything together.