Updates to SD Card logging when motor is on
In order to make the motor controller algorithim run as efficiently as possible, the manner in which SD card writes is done is going to be chagned (only for while the motor is on).  The opening and closing of the file structure is what takes the longest amount of time, and therefor this will only be done when neccessary.  The reason that closing it takes a while is because it clears the buffer on a close.  The buffer is 512 characters long and so it is not neccessary to do this every time.
As a result the file will be opened once upon entering the loop, closed once upon leaving the loop, and the buffer will be cleared after all six of the IMU writes, instead of after every individual write.  This should make each write take about 200us instead of about 3ms.
Added SD Card Writes while the motor is on.
Added code so that it now stores two sets of IMU calibration data in memory.  The first 3 pieces are IMU0 and the second three are IMU1.
Changed the functions in I2C because now we're using a different sensor that has both the temperature and pressure in one sensor.
I2C is going to be completely rewritten not using the Arduino libraries.  Those libraries have too many restrictions, and therefor can not be utilizied for what we need.
This will be part of the next version.