Temperature/Pressure Sensor Read/Write
Added temp/pressure read every second

Added temperature calibration data read on startup

Added pressure calibration read on startup

The writes to the SD card were added.

The writes for temperature and pressure sensors take a really long time due to their size.  As a result the motor on method is used in order to keep it in a reasonable time window. The card is opened once, both values are written, and then the SD card is closed.

Read/Write are also added for when motor is on.