ADC Working
A loop was added so that the motor algorithim is run 8 times in between reads and writes.  This includes a 10ms delay.  This delay can be taken back out if other functionality needs to be added.  This keeps it so that the IMU is only being written to the SD card once every 70ms, which is the desired time.  This will make the motor more precise.
In order to read from all nine battery cells both ADC modules will be used.  The initialization was added to intialize both of them.  The isr was set up for  the second ADC module, and a switch statement is used in both of the isrs.  This is used so that this can be run in the background while other data is being processed.
Added SD Card write of this data once per second.
Added function that checks the level of the batteries.  If they fall below 3.3V the battery is shut down.  If it's the main battery, the backup is turned on and the main is turned off.  If it's the motor battery then the motor is simply disabled.
Once per second the data is written to the SD card.  When this write occurs it kicks off the next series of conversions.  These conversions happen in the background (they're controlled by the isr).  This sequence should only take a few milliseconds, but it happens in the background while other processing is happening in the foreground.
The functionality of the ADC has been tested and verified.