Multiple SD Card Files
A series of SD card files were added.  Because they need to be of type const char* they were hardcoded in an array.  It is set up so that a new file can be added every 5 minutes for 4 hours.  After it adds the last file it will just continue to write to this file.
The file number that we are on is written to non-volitiale memory.  This is done so that if we lose power we don't overwrite the data that is already written
The RTC is checked on startup to see if it is already initialized.  This is done in case the teensy loses power but the RTC does not.  If this is the case the old  RTC values are used.