Current Sensor Added, Miscellaneous Updates
Added a check where the user can press the button in order to restart the file system.  If the button is pressed before the LED starts to flash the calibration on the IMU will be performed.  If the button is pressed in the ~4second window while the LED is flashing rapidly the file system will restart at number 0.  Anytime after this window the button will not do anything.
The file names were shortened from HABIP#.csv to H#.  This was done in order to save on RAM.
Added a function to display the results of the Self test on the LEDs. A boolean array was added to store the results. The array is formatted as follows:
[Current Sensor, ADC0, ADC1, DAQCS SPI, Temp/Press, IMU0, IMU1, SD Card]
There are two kinds of failures defined, "Hard Failures", and "Soft Failures".  Hard failures make it such that the unit is not operational.  A soft failure means that something is not working, but the unit can fly anyway.  A hard failure is defined by two different scenarios:
1) IMU0 and IMU1 Failures
2) SD Card Failure
A soft failure is any other failure that occurs.
A hard failure is defined by the Red LED on the board being turned on and left on until reset.
A soft failure is defined by blinking the Red LED.  The LED is blinked the position in the array + 1 times, ie. Current sensor = 1 time.
If multiple soft failures occur there is a two second delay between each set.
The green LED is turned on if there is no Hard Failure.
I2C interface with the Currrent sensor was written.  It uses the same style of read and writes as the pressure/temperature sensor, and so that code was reused.
Once a second a read is done from the two current sensors and stored on the SD card.  It is just one byte of data.
A ST is added for the current sensor where it reads from the ID register on the board.  This is how it's marked as a pass or a fail.  The sensor is initialized either way and then the data is read.
The Current sensor has not been tested.  It just came in the mail and will be tested in the upcoming weeks.