Added SPI to MSP430
Added SPI Interface with the MSP430.
Messages are parsed and data is formatted to be sent out.  Everything is done inside the ISR, because a request from the MSP430 can come at any point in time.
This was tested using a test bench constructed in a seperate Visual Studio file.  The MSP430 code has not yet been written, and so this was the only method to test.
The functionality of the algorithim was verified, but the functonality of the SPI slave configuration cannot be verified until the MSP430 code has been written.