Miscellanious Updates found in Testing and Integration
Changed the way SD files are named.  const_cast<char*> is now used to convert from a char* to a const char*, that way the file name can be created on the fly.
The new format of file names are hhmmss.csv.  This works as a timestamp
According to the schematic the addresss should be 6F because both pins should be grounded.  It seems like one of the pins is actually not connected, and so the address is actually 0x6E.  If the pin is properly grounded the address should be changed back to 0x6F.
Updated so that IMU1 is the default IMU (the one that's centered around the motor) and if it fails then IMU0 is used.
A retry was added on the calibration because sometimes on come up the data comes back corrupted.  Five retries were added with a 50ms delay in between each retry.
It seems that when the IMU goes from cold to running, the data comes back corrupted on the first attempt.  These short delays seem enough where only a single retry is neccessary to get a good calibration measurement.  The five retries are in there in order for some wiggle room.
The way that the batteries are handled is being slightly changed. During testing it was found that there was a lot of noise on the individual cell lines.  After talking to Carlos it was detemined that this cannot be avoided (he seemed surprised we were only having 150mV on each line, and not more).  His recomendation was to only read the rail voltage and to make the determination based off of that.  So all the cells are still being read and written to the SD card, but to determine if the battery should be turned on or off just the rail voltage is used.  It is also changed so that 10 consecutive good or bad readings need to happen in order for the battery to be turned on or off.