SPI with DAQCS Board
A derivative was added to the controller, making it a PID instead of a PI.  The values are now kp = 120, ki = 0.2, kd = 0.01.  Testing with the current setup (Two PCBs the arduino board, the motor controller, motor and three batteries mounted to the chassis) has shown to be good.  The numbers might need to be adjusted again once everything is connected.
A delay was added after the write to the motor of 50ms.  It was found that writing too fast caused problems.  We were writing at a rate significantly faster than the mechanical time constant, and this was giving issues.  Now with this delay everything seems happy.
The SPI is added but IT IS NOT WORKING.  The algorithim and logic was tested using a script and is working, but something with the configuration is not working.  Due to time constraints this will be left as is.  If time permits this will be looked at, but for not it is not functioning.
A function was added to calculate the pressure on the board.  There is a very good chance that the communications board will not be working in time for the launch.  So Dr. Patru has requested that everything is done locally. Part of this is to cut down at a certain alitude (40,000 ft + 3hrs).  This is approximated using pressure measurements.  So this function has been added.
It has not been called yet, I'm meeting with Carlos tomorrow to discuss how we want to go about this.
NOTE TO 2018-2019 TEAM: THIS IS THE VERSION YOU SHOULD USE AS YOUR ORIGINAL BENCHMARK.  EVERYTHING ADDED AFTER THIS IS PATCHES IN ORDER TO MAKE OUR FLIGHT OPERATIONAL, DUE TO CONSTRAINTS WITH OTHER SUBSYSTEMS.