Updates for Launch
Some of the launch day patches were added.  This includes using the altitude to check for 40,000 feet and 10,000 feet.  It triggers a cutdown at three hours and 40,000 feet.  The motor turns on at either 10,000 feet or ten minutes.  The timer for these two start once we are past two hundred feet from the ground.  the two hundred foot measurement is the only relative measurment.  The other two are taken from sea level.  The original time is stored in flash memory.  It is reset when the button is held on startup.
There is a problem with the RTC losing power.  We thought using a super cap would be enough to keep it going, but that doesn't seem to be enough.  As a result I am using the SD card to restore the RTC.  I put the SD file number in flash memory, and use this to estimate where the RTC was.  A new SD file is created every five minutes, and so this makes it possible to estimate.  TO NEXT YEARS TEAM: Even though last version is the good version, you may want to bring over this RTC functionality.
The RTC is reset if the button is held down on startup.  Right now if the button is held on startup it will calibrate the gyroscope, get a ground pressure measurment, and reset the RTC.